updates to in-text search for debounce on typing, and refresh search index on editing the text
