+++ resources/js/divEditor/saveQueue.js
@@ -14,6 +14,11 @@ import { isPasteOperationActive } from '../paste';
 import { verbose } from '../utilities/logger.js';
 import { clearChunkLoadingInProgress } from '../utilities/chunkLoadingState.js';
 import { markCacheDirty } from '../utilities/cacheState.js';
+import { debounce } from '../utilities/debounce.js';
+import { invalidateSearchIndex } from '../search/inTextSearch/searchToolbar.js';
+
+// Re-export debounce for backwards compatibility
+export { debounce };
 
 // ================================================================
 // DEBOUNCING INFRASTRUCTURE
@@ -28,44 +33,6 @@ const DEBOUNCE_DELAYS = {
   BULK_SAVE: 2000,    // Wait 2s for bulk operations (was 1000ms)
   TITLE_SYNC: 500,
 };
-
-/**
- * Creates a debounced function that delays invoking `func` until after `delay`
- * milliseconds have passed since the last time the debounced function was invoked.
- *
- * Includes `.cancel()` and `.flush()` methods.
- */
-export function debounce(func, delay) {
-  let timeoutId;
-  let lastArgs;
-  let lastThis;
-
-  const debouncedFunction = function (...args) {
-    lastThis = this;
-    lastArgs = args;
-    clearTimeout(timeoutId);
-
-    timeoutId = setTimeout(() => {
-      func.apply(lastThis, lastArgs);
-      timeoutId = null;
-    }, delay);
-  };
-
-  debouncedFunction.cancel = function () {
-    clearTimeout(timeoutId);
-    timeoutId = null;
-  };
-
-  debouncedFunction.flush = function () {
-    if (timeoutId) {
-      clearTimeout(timeoutId);
-      timeoutId = null;
-      func.apply(lastThis, lastArgs);
-    }
-  };
-
-  return debouncedFunction;
-}
 
 // ================================================================
 // SAVE QUEUE CLASS
@@ -159,6 +126,8 @@ export class SaveQueue {
         await batchUpdateIndexedDBRecords(recordsToUpdate);
         // ✅ Mark cache dirty after successful saves
         markCacheDirty();
+        // ✅ Invalidate search index so next search reflects edits
+        invalidateSearchIndex();
       }
 
       if (deletions.length > 0) {
@@ -167,6 +136,8 @@ export class SaveQueue {
         ));
         // ✅ Mark cache dirty after successful deletions
         markCacheDirty();
+        // ✅ Invalidate search index so next search reflects edits
+        invalidateSearchIndex();
       }
 
     } catch (error) {
@@ -213,6 +184,8 @@ export class SaveQueue {
 
       // ✅ Mark cache dirty so it refreshes before next chunk load
       markCacheDirty();
+      // ✅ Invalidate search index so next search reflects deletions
+      invalidateSearchIndex();
     }
   }
